scd/scd: generate device_id.conf also if not in provisioning mode
In case of updates to previous version, where the device uuid was
in the device.conf, it now could happen that there is no device_id.conf.
This would result in startup failures of the cmld. Thus, we now do
not generate the device id only in provisioning mode once, but also
during every start of the scd if it does not exists. In those cases
we extract the uuid from the device.cert which always is available
after the initial provisioning run.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>