scd/usbtoken: Handle SE connection errors while token is plugged
In case, the SE is removed from the usbtoken this was not handled
correctly. Causing errors on which the SE would not be recognized
anymore. Now if the SE is temporarily disconnected (e.g.) by a
power glitch we can recover the connection and every thing is fine.
If the SE is removed for a longer time, we treat this as token
removal and inform the cmld with the lately introduced scd event
SE_REMOVED. The cmld already has the facility to stop the container
in this case.

If the SE is plugged in again, we recover at the next token request
from the cmld.

For the SE removal detection during normal run, we us a watchdog
timer which is currently using queryPIN() to detect if the SE
(SmartCardHSM) is still available in the token.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>