scd/{scd,token}: Helper to get generic scd_token by its internal token
The token uuid is stored in the generic 'scd_token_t'. To make this
available in the concrete token implementations, we need to get through
the list of available scd_tokens and compare the internal token pointer
with the provided pointer to the concrete token implementation, e.g,
usbtoken_t. Since the pointer to the internal concrete token int_token,
is hidden in token.c, we introduce 'token_has_internal_token()' there.
This helper is then called by 'scd_get_token_from_int_token()' in scd.c.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>