scd/control: introduce scd event notifications
Provide a mechanism to notify the cmld with asynchronous events.
We allow one event listener which can be registered by
protobuf DeamonToToken Code REGISTER_EVENT_LISTENER. The scd
responds with REGISTER_EVENT_LISTENER_OK or
REGISTER_EVENT_LISTENER_ERROR in case the global event_fd is already
used.

The first event which can be used is the TOKEN_SE_REMOVED event
which a followup commit will emit on SE removal.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>