daemon/scd: handle scd SE_REMOVED event
We register the first connection of the cmld to the scd as event
listener in scd_init(). Further we register an event handler in
the event loop. This handler handles incoming scd events.
For now this is only SE_REMOVED, on which we stop the container
by calling cmld_container_stop().

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>