scd/usbtoken: use special select APDU instead of queryPIN for SE watchdog
Using queryPIN() would intercept established secure channel
communication of the SE. Thus to check if the SE is still available,
we now use a special 'unselectable' AID (0x00, 0x00, 0x00, 0x00, 0x00)
to which the Card Manager on the SE will answer with 0x6A86 (wrong P1 or
P2 parameter). This won't disturb current Applet selection.

We provide a new helper function usbtoken_is_card_present() which is
used in usbtoken_se_comm_watchdog_cb() instead of queryPIN().

This helper sends the select APDU with processAPDU() function to directly
communicate with the SE and not the reader.

The usually foreseen GET_STATUS APDU to the reader is not handled as
expected by the USB token. Thus, we use this mechanism to check
card presence.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>