daemon/c_hotplug: Deny access to USB tokens
Using the new explicit deny list approach of the c_cgroups_dev
module, access to USB devices which are used as TOKENs by other
containers are now denied during hotplug events. We maintain a
global list of such tokens, which are is maintained during
compartment_new() and compartment_free().

After a plug event occurs it is check if the USB device is in the
global list of tokens. Remember the serial since the minor number of
a USB device is incremented during plug/unplug. If it is, it is
further checked if the container object also is allowed to access the
corresponding low level device. If so, the USB dev is placed in a
list of devices which should be re-allowed after unplug of the token.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>
[felix.wruck@aisec.fraunhofer.de: fixed typo in conditional, skip mknod
for ignored token devices, added TRACE message]
Signed-off-by: Felix Wruck <felix.wruck@aisec.fraunhofer.de>