daemon/c_net: Rename physical interfaces during container cleanup
To avoid name clashes on container tear down, we have to rename the
physical interfaces which where moved to the container. Otherwise,
situations occur where the interfaces was renamed inside the
container clashes with an interface inside the rootns. This may lead
to dangling network namespace, which keeps the corresponding physical
interface forever.

After container init has died we now rename physical interfaces in
the corresponding cleanup() hook. We fork and join the container
netns by the corresponding path in the file system and use a
collision free name utilizing the container_uid() plus a local index
as naming scheme on all physical network interfaces inside the
container.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>