c_seccomp/seccomp: Fix use-after-free on notify fd exception
If the client gets terminated, e.g, by compartment_kill(), the notify
handler will receive an exception in the main process and seccomp->event
will get removed and freed. Afterwards when the seccomp_cleanup()
method is cleaning up the module, it also wants to remove the io event
from event loop and free seccomp->event. This is currently a double free.
Thus, in the handler we just have to set seccomp->event to NULL. This is
checked in compartment_cleanup() and the removal and free is not done
anymore.

This fixes ASAN error:
=================================================================
==464==ERROR: AddressSanitizer: heap-use-after-free on address \
 0x50300001dea0 at pc 0x556f3a0ce5a6 bp 0x7ffea48362c0 sp 0x7ffea48362b0
READ of size 4 at 0x50300001dea0 thread T0
    #0 0x556f3a0ce5a5 in event_remove_io common/event.c:394
    #1 0x556f3a0b80c4 in c_seccomp_cleanup daemon/c_seccomp/seccomp.c:578
    #2 0x556f3a02a854 in compartment_cleanup daemon/compartment.c:656
    #3 0x556f3a02cf4b in compartment_sigchld_cb daemon/compartment.c:746
    #4 0x556f3a0d09d9 in event_signal_handler common/event.c:780
    #5 0x556f3a0d09d9 in event_loop common/event.c:851
    #6 0x556f3a00b12c in main daemon/main.c:146
    #7 0x7f3dee8f5863 in __libc_start_call_main ../sysdeps/nptl/libc_start_call_main.h:58
    #8 0x7f3dee8f590a in __libc_start_main_impl ../csu/libc-start.c:389
    #9 0x556f3a00d3b4 in _start (/usr/sbin/cmld+0x1393b4)

Fixes: 3353cdea2fbf ("daemon/compartment: Introduce c_seccomp module")
Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>