daemon/compartment: Use clone3 instead of clone
To allow future use-cases, e.g., clone flags for new Namespace which
are only possible with clone3. We switch to clone3 syscall now. Also
just use the same stack area as the parent by setting corresponding
clone_args for stack to 'NULL' and stack_size to '0'.
This reduces code complexity.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>