daemon/c_seccomp: introduce sysinfo() emulation
Currently lxcfs is used to virtualize several procfs files to
use cgroup values instead of system wide values for resources.

However, not all user space tools rely on procfs but make use of
the sysinfo() syscall, which again provides system wide values
for resource utilization. Only uptime is already be correctly
handled by the time namespace.

Thus, we have to emulate the sysinfo() call to, e.g., provide
ram limits also for applications which use sysinfo.
Values are used either directly derived from cgroups files or in
case of loads[], we use the proc provided values, which are
emulated by lxcfs if available.

sysinfo() is an unprivileged syscall which does not need root
privileges out side the container. Thus, we can just execute the
emulation code fully in the container context, specially we join
all namespaces except the pidns. Thus, the helper process won't
show up inside the container. Further, we additionally unshare a
new mount namespace in which we mount the cgroup2 fs if the
container has not mounted it or has mounted something else at
'/sys/fs/cgroup'.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>