daemon/common: Log verbosity reduction
This commit adapts the logging verbosity of the cmld and libcommon logs.
To this end, the log level of a number of messages ischanged from DEBUG
to TRACE and overrides of the loglevel in common/network.c and common/verity.c
are removed. Also, log messages of checks in common/macro.h are rephrased.

Signed-off-by: Felix Wruck <felix.wruck@aisec.fraunhofer.de>