daemon: Introduced c_xorg_compat module
Recent Linux kernel 6.9 changed device link of parent devices
in the sysfs. Newer xorg servers can handle this. However, to be
compatible with container payload which includes no recent xorg
server, we introduce this module. It will mount a tmpfs on the
/sys/class/graphics directory and links the included files, so
that it will compatible with old matching logic.

Especially, the link in /sys/class/graphics/fb0 may not include
a pci suffix to get recognized as platform device by legacy xorg
servers. Thus we link to bus/platform instead.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>