common/network: more robust parsing in network_list_link_ns
At first lines of network_list_link_ns does not match with lines
returned in an element of the returned list list_line.
Usually a line of list_line contained all line from one interface.
However, sometimes this got garbled. We now read the output of
the sub process with getline() and create a new list_line element
each time we get an new line with ifindex at the beginning and
if not we just append the string of the existing last element.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>