daemon/c_hotplug: deep copy elements of allow_on_unplug_list
To avoid references to destroyed container objects, use deep copies
of container_usbdev_t elements in tracking list allow_on_unplug_list.
E.g., on container reload the corresponding container is freed and so
is the list of usbdevs of which the references were held in other
container's 'allow_on_unplug_list's.
This use-after-free was triggered if the container was reloaded and
started again.

Fixes: 1dd99f4b3968 ("daemon/c_hotplug: Deny access to USB tokens")
Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>