daemon/cmld: reattach token on container reload
On reload the new container object is created with container_new()
and afterwards the previous object is deleted by a call to
container_free(). Makes sens to first try to create the new updated
container object and only if this succeeds, the old object is
destroyed. However on container_new() the token is attached to the
scd and on container_free() it is removed:

Thus we now have the following behavior:

1) We have the token attached in the scd. The config updated kicks
   in and we get a new container object created by container_new().
   A second attach is made in the scd:
      <INFO>  control.c+221: Token already exists.

2) That is ok, we just ignore that in the scd and the new updated
   container object is created successfully.

3) However afterwards the container_free(<outdated container object>)
   is made and also the scd resources are released by an internal
   call to usbtoken_free() in scd:
      <DEBUG> usbtoken.c+603: Closing CT interface (ctn=0) done.

4) Since both container objects map to the same token in the scd, we
   now have no token attached in the scd anymore.

5) An attempt to do anything token related in the cmld now results
   in that the token is not available:
      <ERROR> control.c+254: No token loaded, unlock failed

Simply fix this by reattaching the token after the container_free()
of the outdated container.

Fixes: c13ef819d042 ("daemon/cmld: cleanup observer handling")
Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>