rattestation/README: Adds additional setup instructions
Following the README instructions installs libprotobuf-c and ibmtss in
/usr/local/lib. While the rattestation tool builds successfully, running
it fails since it can load the shared libraries.

One possible fix would be to change the LD_LIBRARY_PATH, however, that
could create issues for the other tools. Instead, this PR adds a section
in the README about symlinking the libraries to the correct location.

Signed-off-by: Gabriel Loyola Daiqui <gabriel.loyola.daiqui@aisec.fraunhofer.de>
[ michael.weiss ] added commit description from github PR-517
Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>