daemon/c_smartcard.c: do not remove token from scd on smartcard_free()
Do not call c_smartcard_scd_token_remove_block() in smartcard_free().
The token is removed on smartcard_destroy(). Otherwise the scd
disconnects the token, too. And usage in a new config would not work
without manual unplug and plug of the physical token device.

Commit 7ae3e11271c4 "daemon/cmld: reattach token on container reload"
introduced a workaround by calling token_attach() after the
container_free() in cmld.c. However, this is racy and will be obsolete
after this commit.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>