common/file: Fix handling of non-regular files in file_copy
In file_copy, the input file size got calculated  using the min of fstat(2)
and the provided count*bs. For non-regular files like /dev/zero this
lead to wrong behaviour as fstat reports 0 for such files.
Now, for positive counts always count*bs is taken as input file size. In
case a write operation returns 0 it is assumed that no input data is left for
copying.

Signed-off-by: Maximilian Peisl <maximilian.peisl@aisec.fraunhofer.de>