daemon/c_net: Join namespace by pid only as fallback
While there is a race condition when joining the namespace by path,
there is also one when joining it by pid. When a container can not be
stopped gracefully, a kill is issued on the process with the pid defined
in net->container. When this happens, it is not possible to join the
namespace by pid of this container.

By default the namespace is now joined by path, if that does not work it
will be joined by pid.

Signed-off-by: Maximilian Peltzer <mpeltzer@blackned.de>