[WIP] common/cryptfs: dm-integrity switch to internal_hash to support discards
The used stacked approach where dm-crypt is using AEAD cipher on top
of dm-integrity does not support discards for SSDs. Thus, we provide
a new parameter for to allow to setup a new crypt table layout without
stacking. We provide stand-a-lone dm-integrity device on which a
dm-crypt device is created. For the dm-integrity device we use
internal_hash with hmac(sha256) for data integrity and authenticity. On
the dm-crypt device we use aes-xts-plain64 for data confidentiality.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>