common/cryptfs: Fixed key length truncation for crypt only device
The crypto key was truncated to 64 byte key length in hex representation.
This would result in a 256bit aes-xts key which in the end results in an
AES 128 bit key for the actual encryption and decryption. Do not use the
CRYPTFS_FDE_KEY_LEN define which is the length of the key, but the newly
introduced CRYPTO_HEXKEY_LEN which is twice the size for the ascii key.

Fixes: 79fc87debaba ("common/cryptfs: switched authenticated encryption scheme to hmac,xts")
Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>