daemon/cmld: set a random key for the c0 for integrity protection
Instead of just using the DUMMY_KEY we now provide a random key
for the hmac() which is used in the device mapper device set up
by c_vol. If the tpm-bound FDE is used, the corresponding hmac key
for c0 is stored in that protected storage and thus also properly
secured.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>