daemon/c_cgroups_dev: Reload BPF program in starting state
During initial startup usb devices are registered in c_hotplug when the
container is in a starting state. Here the function c_cgroups_dev_assign
is called to add assign and allow rules for cgroup. Because the
container was not yet in state booting, the function quit before loading
the bpf program. This would lead to a state in which the bpf program was
not created and loaded but the device already added to the
global_device_allow_list. Without reconnecting the device, the bpf
program was not loaded.

Loading the bpf program in the starting state fixes this issue.

Signed-off-by: Maximilian Peltzer <mpeltzer@blackned.de>