daemon/{compartment,c_vol}: Introduce COMPARTMENT_ERROR_VOL_CORRUPTED
In case of a corrupted keyfile, the disk encryption key is not set in
start_child_early() handler. We now detect this case and expose this
from c_vol to the compartment module by returning the error code
COMPARTMENT_ERROR_VOL_CORRUPTED. In the compartment this is handled
in the sigchld_early_cb() accordingly and the compartment state is
set to state ZOMBIE.

Further, some error handling and return values are fixed.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>