daemon: moved compartment_module list to container module
Currently all compartment modules are container modules, since
all depend on the container as extension. Thus, we move the
corresponding compartment module list to the container module.
Also the registration method ihas been meeoved and renamed from
compartment_register_module() to
container_register_compartment_module() which globally enables
the corresponding module for the container object.

This will allow us to implement a compartment wrapper similar to
the container module, however, with a different set of submodules.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>