daemon/unit: provide socket on_connect handler for units
If a unit is started and an on_connect handler is set. The on_connect
handler is called if unit reaches COMPARTMENT_STATE_RUNNING.

In detail the procedure is as follows:
In unit_new() a on connect handler and socket name has to be set.
During unit_start, a compartment state observer is registered and if
compartment state reaches running, a inotify handler for the
corresponding socket name is regsitered. If the corresponding socket
in the filesystem is created by the process which was started in this
unit, the socket is connected and the on_connect handler is called.

Further, a new paramter restart is introduced. If restart is set in
unit_new(), the unit will auto restart if it is reaching
COMPARTMENT_STATE_STOPPED.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>