daemon/unit, scd: store data_path in to struct unit
The new attribute data_path is added to the struct unit. The data
path should be used to store the application data inside the unit.
The corresponding API is added in as parameter in unit_new() and
the getter unit_get_data_path().

Adapt user in scd module to add the SCD_TOKEN_DIR as data_path.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>