daemon/u_perm: api to set additional list of devices to be allowed
To also allow additional device initially for a unit. We introduce
a per u_perm whitelist. This could be set externally by the wrapper
unit_device_set_initial_allow() for u_perm_set_initial_allow_dev().
In the u_perm a list_t whitelist is generated if the corresponding
names have a device node mapped in the host.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>