daemon, scd: set device_id.conf to new location
The device_id.cond is now located in scd's data path, aka
SCD_TOKEN_DIR. This is neede for unit start of scd. Otherwise the
scd won't be able to write the device_id.conf. As an update path
with legacy location at <cmld_path>/device_id.conf, we just move
it to SCD_TOKEN_DIR early prior to scd unit is started.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>