daemon/c_hotplug: initially allow usbtoken devce in scd unit
If the token is already plugged at boot the corresponding device
node has to be created initially in the scd_unit, too. Thus, not
only on plug events we also allow access to the scd_unit initially
in c_hotplug_new().

Further, we have to ensure that the initial allow of the hotplug
module is performed before the token is added to the scd by the
c_smartcard module in c_smartcard_new(). Therefore, we have to
reorder the module loading in the Makefile by putting c_hotplug.c
before c_smartcard.c.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>