c_seccomp/sysinfo: fixed missing sysnconf() arround _SC_PAGE_SIZE
During loadavg emulation a buffer for the /proc/loadavg file
was allocated with size _SC_PAGE_SIZE. This however is only an enum
which was accidetially taken directly as size. The intended use was to
use the actual page size which is returned by sysconf(_SC_PAGE_SIZE).
This is fixed now.

Fixes: dd7e20d0bc54 ("daemon/c_seccomp: introduce sysinfo() emulation")
Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>