daemon/c_smartcard: Use unit_get_sock_dir() in path for token control
Since scd is running in a unit the token control sockets are located
in /run/socket/<unit name>/cml-tokencontrol. Hence we need to take
this also into account in c_smartcard by using the unit_get_sockdir()
as well as the SOCKET_PREFIX "cml-" for the construction of the correct
directory name.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>