scd/scd: Limit hw_name and hw_serial value
X.509, specified in RFC 5280, restricts the length of the Common Name (CN) attribute to 64 characters.

In scd.c, the hw_name and hw_serial is combined to the Common Name.
If this combined value exceeds the limit of 64 characters, the cml can't create a CSR.
For example: VMWare exceeds the limits of these characters with it's uuid serial and the name.

This is fixed by a limit in the file_read_new function.
hw_serial is limited to 40 chars in file_read.
hw_name is limited to 20 chars in file_read.

Signed-off-by: ewachter-dev <ewachter@blackned.de>