daemon/guestos_mgr: Check for used kernel images
A/B update requires keeping more than only the latest version of the
kernel guestos in the guestos store. Usually (if everything works fine)
these are the last two versions, however, this may change if an update
was not successful and was rolled back to the previous version. Then the
'oldold' version on the backup must be kept.

Add an additional check that keeps a guestos if one of the redundant
mount options points corresponds to the kernel image of the guestos.

Signed-off-by: Johannes Wiesböck <johannes.wiesboeck@aisec.fraunhofer.de>