daemon/unit: split out socket handling in u_service
The compartment module needs an u_* module with name "c_service"
to handle state COMPARTMENT_STATE_BOOTING, otherwise the compartment
module would directly switch to state RUNNING after state STARTING.
We now also use the state BOOTING for initial startup of the unit
until we can connect to the socket inside the unit. The state
handler using inotify to check if the unit created the socket is
now located in the new u_service module and the u_service module
does the state transition from BOOTING to RUNNING there.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>