daemon/control: include status of units in protobuf responses
LIST_CONTAINERS which assembles a list of UUIDs as response now
uses the optional protobuf field system_services. If set also the
cmld's internal list of units is taken into account.
CONTAINER_STATE also provides unit information if system_services
is set or if a single UUID of a unit is set.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>