daemon/mount: Introduce mount_idmapped()
Refactored idmapped mount generic syscall compat stuff mount module.
Provide a new mount_idmapped header which inludes most of the compat
code.

Use this for a generic mount_idmapped() implementation. This could
later be used in units as well as for a major refactoring in the
c_idmapped module.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>

mount idmapped