daemon/scd: notify containers about (re-)connection of the scd
In case of connection error due to unit restart the scd_on_connect_cb()
is triggered if the scd is available again. Thus notify all container
instances by calling the new container_scd_connect handler. This
reconnects the socket and establishes  a new session to the scd for
the corresponding container instance.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>