common/logf: add condition to the unlink mechanism
If a symlink has a valid and existing target, it will be unlinked
on the daily log roll over or on restart. Then, a new symlink is
created.

If a symlink has no valid / existing target,  the symlink will be
not unlinked.

-> A broken / not valid symlink (.current file) remains and will
not be updated.

This is fixed by extending the condition for unlinking the current
symlink. If the symlink (.current file) is a link, but the target
file does not exist anymore, it will be nevertheless unlinked.

Fixes: 2703ace06ca8 ("common/logf: Creates symlink to identify current log.")
Signed-off-by: Elias Wachter <ewachter@blackned.de>
[michael.weiss: formatted commit message]
Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>