daemon/c_run: avoid double free on readloop error
On massive amount of output data the readloop runs into an error.
This caused a double free in the error handling code of the readloop

The sig child handler of the injected process already handles the
memory deallocation for the corresponding session object. We can
just remove this from the error handling code in the session's
read write loops. This avoids the double free which causes the
cmld to abort after the following log trace:

 <ERROR> c_run.c+488: Readloop returned an error, cleanup!
 <DEBUG> mem.c+130: [MEM] Freeing element 0
 <DEBUG> mem.c+137: [MEM] Freeing array
 <ERROR> c_run.c+488: Readloop returned an error, cleanup!
 <DEBUG> mem.c+137: [MEM] Freeing array

Fixes: 4a5fb1e9e8ad ("daemon/c_run: introduce session for simultaneous running commands")
Reported-by: Harshal Gohel <hg@simonwunderlich.de>
Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>